% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetInput.R
\name{datasetInput}
\alias{datasetInput}
\title{Data input module}
\usage{
datasetInput(id, filter = NULL)
}
\arguments{
\item{id}{module id}

\item{filter}{limit the options of built-in datasets (i.e., \code{is.data.frame} or \code{is.matrix})}
}
\value{
list of data object names
}
\description{
\emph{I use (the \code{filter}) argument to optionally filter the objects found in the datasets package, then create a \code{selectInput()}.}

This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#getting-started-ui-input-server-output}{Getting started: UI input + server output} section of Mastering Shiny.
}
