% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetServer.R
\name{datasetServer}
\alias{datasetServer}
\title{Data server module}
\usage{
datasetServer(id)
}
\arguments{
\item{id}{module id}
}
\value{
reactive data object from datasets package
}
\description{
\emph{we just use \code{get()} to retrieve the dataset with its name...Here we take advantage of the usual rule that last expression processed in the function becomes the return value. This value should always be a reactive}

This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#getting-started-ui-input-server-output}{Getting started: UI input + server output} section of Mastering Shiny.
}
