% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDataVarApp.R
\name{DataVarUI}
\alias{DataVarUI}
\title{Data and variable shiny UI}
\usage{
DataVarUI()
}
\value{
shiny UI function
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#modules-inside-of-modules}{Modules inside of modules} section of Mastering Shiny.
}
