% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_find_vars.R
\name{find_data_vars}
\alias{find_data_vars}
\title{Find variables (improved)}
\usage{
find_data_vars(data, filter)
}
\arguments{
\item{data}{reactive data object to retrieve variables from}

\item{filter}{which variables to list. \emph{This will be set by the caller of the module, so doesn’t need to be reactive}}
}
\value{
variable names
}
\description{
This is the improved \code{find_vars()} function.

\emph{check that each input to your module is either reactive or constant. If you don’t, and the user supplies the wrong type, they’ll get a cryptic error message. You can make the life of module user much easier with a quick and dirty call to \code{stopifnot()}}

This function comes from the \href{https://mastering-shiny.org/scaling-modules.html#server-inputs}{Server inputs} section of Mastering Shiny.
}
