% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_find_vars.R
\name{find_vars}
\alias{find_vars}
\title{Find variables}
\usage{
find_vars(data, filter)
}
\arguments{
\item{data}{reactive data object to retrieve variables from}

\item{filter}{which variables to list. \emph{This will be set by the caller of the module, so doesn’t need to be reactive}}
}
\value{
variable names
}
\description{
This function comes from the \href{https://mastering-shiny.org/scaling-modules.html#case-study-selecting-a-numeric-variable}{Case study: selecting a numeric variable} section of Mastering Shiny.
}
