% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_datameta}
\alias{get_pkg_datameta}
\title{Get package datasets metadata}
\usage{
get_pkg_datameta(
  package,
  allClass = FALSE,
  incPackage = length(package) > 1,
  maxTitle = NULL
)
}
\arguments{
\item{package}{name of package (a character vector)}

\item{allClass}{logical (include all classes of data?)}

\item{incPackage}{logical (include package name in result?)}

\item{maxTitle}{maximum length of dataset title}
}
\value{
\code{data.frame} with 6 variables (\code{dataset}, \code{title}, \code{dimensions},
\code{obs}, \code{vars}, \code{display_title})
}
\description{
This is a variation on the \code{vcdExtra::datasets()} function.
Read more here:
https://github.com/friendly/vcdExtra/blob/master/R/datasets.R
}
\examples{
require(tidyr)
get_pkg_datameta("tidyr")
}
