% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_plot_inputs_list}
\alias{get_plot_inputs_list}
\title{Create plot_inputs() list (for debugging and testing)}
\usage{
get_plot_inputs_list(pkg, ds)
}
\value{
list of plot inputs
}
\description{
Create plot_inputs() list (for debugging and testing)
}
\examples{
get_plot_inputs_list(pkg = "palmerpenguins", ds = "penguins")
}
