% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grp_var_sum}
\alias{grp_var_sum}
\title{Grouped variable summary}
\usage{
grp_var_sum(data, grp_var, var)
}
\arguments{
\item{data}{tibble (unquoted)}

\item{grp_var}{group variable (as quoted string)}

\item{var}{variable to summarize (as quoted string)}
}
\value{
tibble with summary statistics
}
\description{
Grouped variable summary
}
\examples{
require(palmerpenguins)
grp_var_sum(
data = palmerpenguins::penguins,
grp_var = "island",
var = "bill_length_mm")
}
