% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_ui_inputs}
\alias{make_ui_inputs}
\title{Make UI inputs}
\usage{
make_ui_inputs(app_data)
}
\arguments{
\item{app_data}{dataset for app (\code{data.frame} or \code{tibble})}
}
\value{
zeallot assignment (\verb{\%<-\%}) with input character vector on LHS and
list of names by type on the RHS
}
\description{
This is meant to be used in the console--it generates the code for assigning
the elements from a list into a collection of vectors.
}
\examples{
require(palmerpenguins)
make_ui_inputs(palmerpenguins::penguins)
require(zeallot)
c(double_vars, integer_vars, factor_vars) \%<-\%
  list(
    double_vars = c("bill_length_mm", "bill_depth_mm"),
    integer_vars = c("flipper_length_mm", "body_mass_g", "year"),
    factor_vars = c("species", "island", "sex")
  )
double_vars
integer_vars
factor_vars
}
