% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstsap-package.R
\docType{package}
\name{mstsap-package}
\alias{mstsap}
\alias{mstsap-package}
\title{mstsap: Test Your Shiny Package App server functions}
\description{
This shiny app-package if from Mastering Shiny and was developed to demonstrate testing with testthat and shiny's testServer().
}
\keyword{internal}
