% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDataVarApp.R
\name{selectDataVarApp2}
\alias{selectDataVarApp2}
\title{Data and variable shiny app (version 2)}
\usage{
selectDataVarApp2()
}
\value{
shiny app
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#modules-inside-of-modules}{Modules inside of modules} section of Mastering Shiny.
}
