% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDataVarServer.R
\name{selectDataVarServer}
\alias{selectDataVarServer}
\title{Data and variable input server module}
\usage{
selectDataVarServer(id, filter = is.numeric)
}
\arguments{
\item{id}{module id}

\item{filter}{limit the options of built-in datasets (i.e., \code{is.data.frame} or \code{is.matrix})}
}
\value{
shiny server function
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#modules-inside-of-modules}{Modules inside of modules} section of Mastering Shiny.
}
