% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDataVarUI.R
\name{selectDataVarUI}
\alias{selectDataVarUI}
\title{Data and variable input UI module}
\usage{
selectDataVarUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny UI function
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#modules-inside-of-modules}{Modules inside of modules} section of Mastering Shiny.
}
