% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVarInput.R
\name{selectVarInput}
\alias{selectVarInput}
\title{Variable input UI module}
\usage{
selectVarInput(id)
}
\arguments{
\item{id}{module id}
}
\value{
module UI variable input
}
\description{
"\emph{allows the user to select variables of specified type from a given reactive dataset.}"

This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#case-study-selecting-a-numeric-variable}{Case study: selecting a numeric variable} section of Mastering Shiny.
}
