% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVarServer.R
\name{selectVarServer}
\alias{selectVarServer}
\title{Variable input server module}
\usage{
selectVarServer(id, data, filter = is.numeric)
}
\arguments{
\item{id}{module id}

\item{data}{reactive data object to retrieve variables from}

\item{filter}{passed to \code{find_vars()}, i.e., the variables to list. \emph{This will be set by the caller of the module, so doesn’t need to be reactive}}
}
\value{
shiny server function
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#server-inputs}{Server inputs} section of Mastering Shiny.
}
