% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetApp.R
\name{datasetApp}
\alias{datasetApp}
\title{Dataset app}
\usage{
datasetApp(filter = NULL)
}
\arguments{
\item{filter}{limit the options of built-in datasets (i.e., \code{is.data.frame}
or \code{is.matrix})}
}
\value{
shiny application
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#getting-started-ui-input-server-output}{Getting started: UI input + server output} section of Mastering Shiny.
}
