% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDataVarApp.R
\name{selectDataVarApp}
\alias{selectDataVarApp}
\title{Data and variable shiny app}
\usage{
selectDataVarApp(filter = is.numeric)
}
\arguments{
\item{filter}{limit the options of built-in datasets (i.e., \code{is.data.frame} or \code{is.matrix})}
}
\value{
shiny app
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#modules-inside-of-modules}{Modules inside of modules} section of Mastering Shiny.
}
