% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVarServer.R
\name{selectVarServer}
\alias{selectVarServer}
\title{Variable input server module}
\usage{
selectVarServer(id, data, filter = is.numeric)
}
\arguments{
\item{id}{module id}

\item{data}{reactive data object to retrieve variables from}

\item{filter}{passed to \code{find_vars()}, i.e., the type of variables to list (\emph{this will be set by the caller of the module, so doesn’t need to be reactive}}
}
\value{
shiny server function
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#server-inputs}{Server inputs} section of Mastering Shiny.
}
\section{stopifnot()}{


\emph{check that each input to your module is either reactive or constant. If you don’t, and the user supplies the wrong type, they’ll get a cryptic error message. You can make the life of module user much easier with a quick and dirty call to \code{stopifnot()}}
}

